retailmind-agent/requirements.txt
streamlit>=1.32.0
openai>=1.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
plotly>=5.18.0
